Move common schema parser code from SchemaParser into new AbstractSchemaParser abstract class.
